use TESTNET build flag for cli integration tests
