+++ x/crosschain/client/cli_query_tests/cctx.go
@@ -1,4 +1,4 @@
-package cli_query
+package cli_query_tests
 
 import (
 	"fmt"
@@ -11,9 +11,68 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-func (s *CliTestSuite) TestShowGasPrice() {
+func (s *CliTestSuite) TestListCCTX() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.GasPriceList
+	objs := s.state.CrossChainTxs
+	request := func(next []byte, offset, limit uint64, total bool) []string {
+		args := []string{
+			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
+		}
+		if next == nil {
+			args = append(args, fmt.Sprintf("--%s=%d", flags.FlagOffset, offset))
+		} else {
+			args = append(args, fmt.Sprintf("--%s=%s", flags.FlagPageKey, next))
+		}
+		args = append(args, fmt.Sprintf("--%s=%d", flags.FlagLimit, limit))
+		if total {
+			args = append(args, fmt.Sprintf("--%s", flags.FlagCountTotal))
+		}
+		return args
+	}
+
+	s.Run("ByOffset", func() {
+		step := 2
+		for i := 0; i < len(objs); i += step {
+			args := request(nil, uint64(i), uint64(step), false)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListSend(), args)
+			s.Require().NoError(err)
+			var resp types.QueryAllCctxResponse
+			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+			for j := i; j < len(objs) && j < i+step; j++ {
+				s.Require().Equal(objs[j], resp.CrossChainTx[j-i])
+			}
+		}
+	})
+	s.Run("ByKey", func() {
+		step := 2
+		var next []byte
+		for i := 0; i < len(objs); i += step {
+			args := request(next, 0, uint64(step), false)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListSend(), args)
+			s.Require().NoError(err)
+			var resp types.QueryAllCctxResponse
+			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+			for j := i; j < len(objs) && j < i+step; j++ {
+				s.Assert().Equal(objs[j], resp.CrossChainTx[j-i])
+			}
+			next = resp.Pagination.NextKey
+		}
+	})
+	s.Run("Total", func() {
+		args := request(nil, 0, uint64(len(objs)), true)
+		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListSend(), args)
+		s.Require().NoError(err)
+		var resp types.QueryAllCctxResponse
+		s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+		s.Require().NoError(err)
+		s.Require().Equal(len(objs), int(resp.Pagination.Total))
+		s.Require().Equal(objs, resp.CrossChainTx)
+	})
+}
+
+func (s *CliTestSuite) TestShowSend() {
+	ctx := s.network.Validators[0].ClientCtx
+	objs := s.state.CrossChainTxs
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
@@ -22,7 +81,7 @@ func (s *CliTestSuite) TestShowGasPrice() {
 		id   string
 		args []string
 		err  error
-		obj  *types.GasPrice
+		obj  *types.CrossChainTx
 	}{
 		{
 			desc: "found",
@@ -41,76 +100,18 @@ func (s *CliTestSuite) TestShowGasPrice() {
 		s.Run(tc.desc, func() {
 			args := []string{tc.id}
 			args = append(args, tc.args...)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowGasPrice(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowSend(), args)
 			if tc.err != nil {
 				stat, ok := status.FromError(tc.err)
 				s.Require().True(ok)
 				s.Require().ErrorIs(stat.Err(), tc.err)
 			} else {
 				s.Require().NoError(err)
-				var resp types.QueryGetGasPriceResponse
+				var resp types.QueryGetCctxResponse
 				s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-				s.Require().NotNil(resp.GasPrice)
-				s.Require().Equal(tc.obj, resp.GasPrice)
+				s.Require().NotNil(resp.CrossChainTx)
+				s.Require().Equal(tc.obj, resp.CrossChainTx)
 			}
 		})
 	}
-}
-
-func (s *CliTestSuite) TestListGasPrice() {
-	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.GasPriceList
-	request := func(next []byte, offset, limit uint64, total bool) []string {
-		args := []string{
-			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-		}
-		if next == nil {
-			args = append(args, fmt.Sprintf("--%s=%d", flags.FlagOffset, offset))
-		} else {
-			args = append(args, fmt.Sprintf("--%s=%s", flags.FlagPageKey, next))
-		}
-		args = append(args, fmt.Sprintf("--%s=%d", flags.FlagLimit, limit))
-		if total {
-			args = append(args, fmt.Sprintf("--%s", flags.FlagCountTotal))
-		}
-		return args
-	}
-	s.Run("ByOffset", func() {
-		step := 2
-		for i := 0; i < len(objs); i += step {
-			args := request(nil, uint64(i), uint64(step), false)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListGasPrice(), args)
-			s.Require().NoError(err)
-			var resp types.QueryAllGasPriceResponse
-			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			for j := i; j < len(objs) && j < i+step; j++ {
-				s.Assert().Equal(objs[j], resp.GasPrice[j-i])
-			}
-		}
-	})
-	s.Run("ByKey", func() {
-		step := 2
-		var next []byte
-		for i := 0; i < len(objs); i += step {
-			args := request(next, 0, uint64(step), false)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListGasPrice(), args)
-			s.Require().NoError(err)
-			var resp types.QueryAllGasPriceResponse
-			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			for j := i; j < len(objs) && j < i+step; j++ {
-				s.Assert().Equal(objs[j], resp.GasPrice[j-i])
-			}
-			next = resp.Pagination.NextKey
-		}
-	})
-	s.Run("Total", func() {
-		args := request(nil, 0, uint64(len(objs)), true)
-		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListGasPrice(), args)
-		s.Require().NoError(err)
-		var resp types.QueryAllGasPriceResponse
-		s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-		s.Require().NoError(err)
-		s.Require().Equal(len(objs), int(resp.Pagination.Total))
-		s.Require().Equal(objs, resp.GasPrice)
-	})
 }
