+++ x/crosschain/client/cli_query_tests/intx_hash.go
@@ -1,52 +1,22 @@
-package cli_test
+package cli_query_tests
 
 import (
 	"fmt"
-	"github.com/zeta-chain/zetacore/app"
-	"strconv"
-	"testing"
-
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
-	"github.com/stretchr/testify/require"
-	tmcli "github.com/tendermint/tendermint/libs/cli"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 
-	"github.com/zeta-chain/zetacore/testutil/network"
+	tmcli "github.com/tendermint/tendermint/libs/cli"
 	"github.com/zeta-chain/zetacore/testutil/nullify"
 	"github.com/zeta-chain/zetacore/x/crosschain/client/cli"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
+	"google.golang.org/grpc/status"
+	"strconv"
 )
 
-// Prevent strconv unused error
-var _ = strconv.IntSize
-
-func networkWithInTxHashToCctxObjects(t *testing.T, n int) (*network.Network, []types.InTxHashToCctx) {
-	t.Helper()
-	cfg := network.DefaultConfig()
-	state := types.GenesisState{}
-	require.NoError(t, cfg.Codec.UnmarshalJSON(cfg.GenesisState[types.ModuleName], &state))
-
-	for i := 0; i < n; i++ {
-		inTxHashToCctx := types.InTxHashToCctx{
-			InTxHash: strconv.Itoa(i),
-		}
-		nullify.Fill(&inTxHashToCctx)
-		state.InTxHashToCctxList = append(state.InTxHashToCctxList, inTxHashToCctx)
-	}
-	buf, err := cfg.Codec.MarshalJSON(&state)
-	require.NoError(t, err)
-	cfg.GenesisState[types.ModuleName] = buf
-	cfg.GenesisState = network.SetupZetaGenesisState(t, cfg.GenesisState, cfg.Codec)
-	net, err := network.New(t, app.NodeDir, cfg)
-	return net, state.InTxHashToCctxList
-}
-
-func TestShowInTxHashToCctx(t *testing.T) {
-	net, objs := networkWithInTxHashToCctxObjects(t, 2)
-
-	ctx := net.Validators[0].ClientCtx
+func (s *CliTestSuite) TestShowInTxHashToCctx() {
+	ctx := s.network.Validators[0].ClientCtx
+	objs := s.state.InTxHashToCctxList
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
@@ -73,7 +43,7 @@ func TestShowInTxHashToCctx(t *testing.T) {
 			err:  status.Error(codes.NotFound, "not found"),
 		},
 	} {
-		t.Run(tc.desc, func(t *testing.T) {
+		s.Run(tc.desc, func() {
 			args := []string{
 				tc.idInTxHash,
 			}
@@ -81,15 +51,14 @@ func TestShowInTxHashToCctx(t *testing.T) {
 			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowInTxHashToCctx(), args)
 			if tc.err != nil {
 				stat, ok := status.FromError(tc.err)
-				require.True(t, ok)
-				require.ErrorIs(t, stat.Err(), tc.err)
+				s.Require().True(ok)
+				s.Require().ErrorIs(stat.Err(), tc.err)
 			} else {
-				require.NoError(t, err)
+				s.Require().NoError(err)
 				var resp types.QueryGetInTxHashToCctxResponse
-				require.NoError(t, net.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-				require.NotNil(t, resp.InTxHashToCctx)
-				require.Equal(t,
-					nullify.Fill(&tc.obj),
+				s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+				s.Require().NotNil(resp.InTxHashToCctx)
+				s.Require().Equal(nullify.Fill(&tc.obj),
 					nullify.Fill(&resp.InTxHashToCctx),
 				)
 			}
@@ -97,10 +66,10 @@ func TestShowInTxHashToCctx(t *testing.T) {
 	}
 }
 
-func TestListInTxHashToCctx(t *testing.T) {
-	net, objs := networkWithInTxHashToCctxObjects(t, 5)
-
-	ctx := net.Validators[0].ClientCtx
+func (s *CliTestSuite) TestListInTxHashToCctx() {
+	ctx := s.network.Validators[0].ClientCtx
+	objs := s.state.InTxHashToCctxList
+	cctxCount := len(s.state.CrossChainTxs)
 	request := func(next []byte, offset, limit uint64, total bool) []string {
 		args := []string{
 			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
@@ -116,48 +85,47 @@ func TestListInTxHashToCctx(t *testing.T) {
 		}
 		return args
 	}
-	t.Run("ByOffset", func(t *testing.T) {
+	s.Run("ByOffset", func() {
 		step := 2
 		for i := 0; i < len(objs); i += step {
 			args := request(nil, uint64(i), uint64(step), false)
 			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListInTxHashToCctx(), args)
-			require.NoError(t, err)
+			s.Require().NoError(err)
 			var resp types.QueryAllInTxHashToCctxResponse
-			require.NoError(t, net.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			require.LessOrEqual(t, len(resp.InTxHashToCctx), step)
-			require.Subset(t,
-				nullify.Fill(objs),
+			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+			s.Require().LessOrEqual(len(resp.InTxHashToCctx), step)
+			s.Require().Subset(nullify.Fill(objs),
 				nullify.Fill(resp.InTxHashToCctx),
 			)
 		}
 	})
-	t.Run("ByKey", func(t *testing.T) {
+	s.Run("ByKey", func() {
 		step := 2
 		var next []byte
 		for i := 0; i < len(objs); i += step {
 			args := request(next, 0, uint64(step), false)
 			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListInTxHashToCctx(), args)
-			require.NoError(t, err)
+			s.Require().NoError(err)
 			var resp types.QueryAllInTxHashToCctxResponse
-			require.NoError(t, net.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			require.LessOrEqual(t, len(resp.InTxHashToCctx), step)
-			require.Subset(t,
-				nullify.Fill(objs),
+			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+			s.Require().LessOrEqual(len(resp.InTxHashToCctx), step)
+			s.Require().Subset(nullify.Fill(objs),
 				nullify.Fill(resp.InTxHashToCctx),
 			)
 			next = resp.Pagination.NextKey
 		}
 	})
-	t.Run("Total", func(t *testing.T) {
+	s.Run("Total", func() {
 		args := request(nil, 0, uint64(len(objs)), true)
 		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListInTxHashToCctx(), args)
-		require.NoError(t, err)
+		s.Require().NoError(err)
 		var resp types.QueryAllInTxHashToCctxResponse
-		require.NoError(t, net.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-		require.NoError(t, err)
-		require.Equal(t, len(objs), int(resp.Pagination.Total))
-		require.ElementsMatch(t,
-			nullify.Fill(objs),
+		fmt.Println(out.String())
+		s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+		s.Require().NoError(err)
+		// saving CCTX also adds a new mapping
+		s.Require().Equal(len(objs)+cctxCount, int(resp.Pagination.Total))
+		s.Require().ElementsMatch(nullify.Fill(objs),
 			nullify.Fill(resp.InTxHashToCctx),
 		)
 	})
