+++ x/crosschain/client/cli_query_tests/last_block_height.go
@@ -1,4 +1,4 @@
-package cli_query
+package cli_query_tests
 
 import (
 	"fmt"
@@ -11,68 +11,9 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-func (s *CliTestSuite) TestListCCTX() {
+func (s *CliTestSuite) TestShowLastBlockHeight() {
 	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.CrossChainTxs
-	request := func(next []byte, offset, limit uint64, total bool) []string {
-		args := []string{
-			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-		}
-		if next == nil {
-			args = append(args, fmt.Sprintf("--%s=%d", flags.FlagOffset, offset))
-		} else {
-			args = append(args, fmt.Sprintf("--%s=%s", flags.FlagPageKey, next))
-		}
-		args = append(args, fmt.Sprintf("--%s=%d", flags.FlagLimit, limit))
-		if total {
-			args = append(args, fmt.Sprintf("--%s", flags.FlagCountTotal))
-		}
-		return args
-	}
-
-	s.Run("ByOffset", func() {
-		step := 2
-		for i := 0; i < len(objs); i += step {
-			args := request(nil, uint64(i), uint64(step), false)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListSend(), args)
-			s.Require().NoError(err)
-			var resp types.QueryAllCctxResponse
-			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			for j := i; j < len(objs) && j < i+step; j++ {
-				s.Require().Equal(objs[j], resp.CrossChainTx[j-i])
-			}
-		}
-	})
-	s.Run("ByKey", func() {
-		step := 2
-		var next []byte
-		for i := 0; i < len(objs); i += step {
-			args := request(next, 0, uint64(step), false)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListSend(), args)
-			s.Require().NoError(err)
-			var resp types.QueryAllCctxResponse
-			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-			for j := i; j < len(objs) && j < i+step; j++ {
-				s.Assert().Equal(objs[j], resp.CrossChainTx[j-i])
-			}
-			next = resp.Pagination.NextKey
-		}
-	})
-	s.Run("Total", func() {
-		args := request(nil, 0, uint64(len(objs)), true)
-		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListSend(), args)
-		s.Require().NoError(err)
-		var resp types.QueryAllCctxResponse
-		s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-		s.Require().NoError(err)
-		s.Require().Equal(len(objs), int(resp.Pagination.Total))
-		s.Require().Equal(objs, resp.CrossChainTx)
-	})
-}
-
-func (s *CliTestSuite) TestShowSend() {
-	ctx := s.network.Validators[0].ClientCtx
-	objs := s.state.CrossChainTxs
+	objs := s.state.LastBlockHeightList
 	common := []string{
 		fmt.Sprintf("--%s=json", tmcli.OutputFlag),
 	}
@@ -81,7 +22,7 @@ func (s *CliTestSuite) TestShowSend() {
 		id   string
 		args []string
 		err  error
-		obj  *types.CrossChainTx
+		obj  *types.LastBlockHeight
 	}{
 		{
 			desc: "found",
@@ -100,18 +41,76 @@ func (s *CliTestSuite) TestShowSend() {
 		s.Run(tc.desc, func() {
 			args := []string{tc.id}
 			args = append(args, tc.args...)
-			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowSend(), args)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdShowLastBlockHeight(), args)
 			if tc.err != nil {
 				stat, ok := status.FromError(tc.err)
 				s.Require().True(ok)
 				s.Require().ErrorIs(stat.Err(), tc.err)
 			} else {
 				s.Require().NoError(err)
-				var resp types.QueryGetCctxResponse
+				var resp types.QueryGetLastBlockHeightResponse
 				s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
-				s.Require().NotNil(resp.CrossChainTx)
-				s.Require().Equal(tc.obj, resp.CrossChainTx)
+				s.Require().NotNil(resp.LastBlockHeight)
+				s.Require().Equal(tc.obj, resp.LastBlockHeight)
 			}
 		})
 	}
+}
+
+func (s *CliTestSuite) TestListLastBlockHeight() {
+	ctx := s.network.Validators[0].ClientCtx
+	objs := s.state.LastBlockHeightList
+	request := func(next []byte, offset, limit uint64, total bool) []string {
+		args := []string{
+			fmt.Sprintf("--%s=json", tmcli.OutputFlag),
+		}
+		if next == nil {
+			args = append(args, fmt.Sprintf("--%s=%d", flags.FlagOffset, offset))
+		} else {
+			args = append(args, fmt.Sprintf("--%s=%s", flags.FlagPageKey, next))
+		}
+		args = append(args, fmt.Sprintf("--%s=%d", flags.FlagLimit, limit))
+		if total {
+			args = append(args, fmt.Sprintf("--%s", flags.FlagCountTotal))
+		}
+		return args
+	}
+	s.Run("ByOffset", func() {
+		step := 2
+		for i := 0; i < len(objs); i += step {
+			args := request(nil, uint64(i), uint64(step), false)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListLastBlockHeight(), args)
+			s.Require().NoError(err)
+			var resp types.QueryAllLastBlockHeightResponse
+			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+			for j := i; j < len(objs) && j < i+step; j++ {
+				s.Assert().Equal(objs[j], resp.LastBlockHeight[j-i])
+			}
+		}
+	})
+	s.Run("ByKey", func() {
+		step := 2
+		var next []byte
+		for i := 0; i < len(objs); i += step {
+			args := request(next, 0, uint64(step), false)
+			out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListLastBlockHeight(), args)
+			s.Require().NoError(err)
+			var resp types.QueryAllLastBlockHeightResponse
+			s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+			for j := i; j < len(objs) && j < i+step; j++ {
+				s.Assert().Equal(objs[j], resp.LastBlockHeight[j-i])
+			}
+			next = resp.Pagination.NextKey
+		}
+	})
+	s.Run("Total", func() {
+		args := request(nil, 0, uint64(len(objs)), true)
+		out, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdListLastBlockHeight(), args)
+		s.Require().NoError(err)
+		var resp types.QueryAllLastBlockHeightResponse
+		s.Require().NoError(s.network.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
+		s.Require().NoError(err)
+		s.Require().Equal(len(objs), int(resp.Pagination.Total))
+		s.Require().Equal(objs, resp.LastBlockHeight)
+	})
 }
